Refactor: 칸반 카드 컴포넌트 리팩토링 #39
